adapters: cut vestigial shims, consolidate duplicates, dead-code + correctness (#23/#24/#26)
#23 — _room_facade.py: remove unreachable LocalParticipant shims (register_rpc_method,
  unregister_rpc_method, set_track_subscription_permissions, send_file; verified zero
  callers). Kept perform_rpc (a test asserts NotImplementedError) and the live-path
  publish_transcription/stream_text/stream_bytes. Dropped dead _token/_server_url +
  now-unused imports. Consolidated the near-verbatim duplicated SoundfileMixer and
  AudioRecorder into shared agent_transport/_pipecat_{mixers,processors}.py; the four
  per-package modules are now thin re-exports (both import paths resolve to one class).
  Fixed stale _StubJobContext references in examples.

#24 — removed dead frame-type imports left by the InterruptionFrame change
  (audio_stream_transport.py: InterruptionFrame/StopFrame/OutputTransportMessage*;
  sip_transport.py keeps InterruptionFrame, used by the restored local clear),
  dead os/tempfile/StartFrame in processors, dead 'import json' in server.py; fixed
  the now-inaccurate module docstrings.

#26 — _room_facade.py: emit 'disconnected' once via a _disconnected_emitted guard
  (was double-emitted across disconnect()/_on_session_ended()). server.py: dedup
  participant_connected synchronously by reserving an _active_calls placeholder before
  create_task (mirrors _outbound_session_ids), so a duplicate event can't spawn a
  second call. AudioRecorder: explicit _stop_fired bool reset in start_recording so a
  recorder can restart (was a never-cleared hasattr guard).

NOT changed: init_recording's live-dict mutate+restore — copying instead would defeat
  AgentSession's RecorderIO suppression and double-record; fixed the misleading docstring.

pyflakes clean; pytest 162 passed.
