+++ crates/agent-transport-python/adapters/agent_transport/sip/pipecat/processors.py
@@ -1,95 +1,11 @@
 """Rust-backed pipeline processors for Pipecat SIP transport.
 
-AudioRecorder hooks Rust's transport-level recording into the Pipecat pipeline.
-
-Usage:
-    from agent_transport.sip.pipecat import AudioRecorder
-
-    recorder = AudioRecorder(transport, path=f"/tmp/agent-sessions/recording_{transport.session_id}.ogg", num_channels=2)
-    pipeline = Pipeline([transport.input(), stt, llm, tts, transport.output(), recorder])
+Thin re-export of the shared, transport-agnostic implementation in
+``agent_transport._pipecat_processors``. Kept so existing imports
+(``from agent_transport.sip.pipecat import AudioRecorder`` /
+``from .processors import AudioRecorder``) keep working.
 """
 
-from typing import Optional
+from agent_transport._pipecat_processors import AudioRecorder
 
-from loguru import logger
-
-try:
-    from pipecat.frames.frames import CancelFrame, EndFrame, Frame, StartFrame
-    from pipecat.processors.audio.audio_buffer_processor import AudioBufferProcessor
-    from pipecat.processors.frame_processor import FrameDirection
-except ImportError:
-    raise ImportError("pipecat-ai is required: pip install pipecat-ai")
-
-
-class AudioRecorder(AudioBufferProcessor):
-    """AudioBufferProcessor + Rust file recording for SIP calls.
-
-    Inherits all AudioBufferProcessor behavior:
-    - on_audio_data, on_track_audio_data, on_user_turn_audio_data, on_bot_turn_audio_data
-
-    Adds Rust transport-level file recording:
-    - Records directly in Rust's RTP send loop (zero Python overhead)
-    - OGG/Opus output, stereo (L=user, R=agent)
-    - on_recording_stopped(path) event when file is written
-
-    If path is not provided, behaves exactly like AudioBufferProcessor.
-    """
-
-    def __init__(
-        self,
-        transport: "SipTransport",
-        *,
-        path: Optional[str] = None,
-        stereo: bool = True,
-        sample_rate: Optional[int] = None,
-        num_channels: int = 1,
-        buffer_size: int = 0,
-        enable_turn_audio: bool = False,
-        **kwargs,
-    ):
-        super().__init__(
-            sample_rate=sample_rate,
-            num_channels=num_channels,
-            buffer_size=buffer_size,
-            enable_turn_audio=enable_turn_audio,
-            **kwargs,
-        )
-        self._transport = transport
-        self._path = path
-        self._stereo = stereo
-        self._rust_recording = False
-        self._register_event_handler("on_recording_stopped")
-
-    async def start_recording(self):
-        """Start recording. Starts both Python buffering and Rust file recording."""
-        await super().start_recording()
-        if self._path:
-            try:
-                self._transport.start_recording(self._path, self._stereo)
-                self._rust_recording = True
-            except Exception as e:
-                logger.warning("Rust recording failed to start: {}", e)
-
-    async def stop_recording(self):
-        """Stop recording. Stops Rust file, then fires Python callbacks."""
-        if not self._path or hasattr(self, '_stop_fired'):
-            return
-        self._stop_fired = True
-        if self._rust_recording:
-            try:
-                self._transport.stop_recording()
-            except Exception as e:
-                logger.warning("Rust recording failed to stop: {}", e)
-            self._rust_recording = False
-        await super().stop_recording()
-        await self._call_event_handler("on_recording_stopped", self._path)
-
-    async def process_frame(self, frame: Frame, direction: FrameDirection):
-        if isinstance(frame, (EndFrame, CancelFrame)) and self._rust_recording:
-            try:
-                self._transport.stop_recording()
-            except Exception as e:
-                logger.warning("AudioRecorder stop_recording failed: {}", e)
-            self._rust_recording = False
-
-        await super().process_frame(frame, direction)
+__all__ = ["AudioRecorder"]
