+++ examples/livekit/sip_agent.py
@@ -77,7 +77,8 @@ class Assistant(Agent):
             # - send_dtmf_events: works via Room facade
             # - EndCallTool: beta tool that ends the call when the user says
             #   goodbye. Hooks into job_ctx.shutdown() + job_ctx.delete_room(),
-            #   both of which we implement on _StubJobContext.
+            #   both of which the JobContext implements (via
+            #   TransportJobContextMixin) so the call is dropped cleanly.
             tools=[send_dtmf_events, EndCallTool()],
         )
